Add check that user is logged in on private calls

+++ admin/inc/portfolioFunctions.php
@@ -6,6 +6,19 @@
 
 date_default_timezone_set("Europe/London");
 
+function isLoggedIn() {
+	if (!isset($_SESSION)) {
+		session_start();
+	}
+
+	if (!empty($_SESSION["username"]) && $_SESSION["username"] === PORTFOLIOUSERNAME
+			&& !empty($_SESSION["password"]) && $_SESSION["password"] === PORTFOLIOPASSWORD) {
+		return true;
+	}
+
+	return false;
+}
+
 //gets a user, either trying to log in or trying to get information of a user
 function login($data)
 {
@@ -141,40 +154,47 @@ function getProjects($data)
 //add a project user has attempted to add
 function addProject($data)
 {
+	//checks if user is authored
+	if (isLoggedIn()) {
 
-	//checks if requests needed are present and not empty
-	$dataNeeded = array("projectName", "skills", "longDescription", "shortDescription", "github", "date");
-	if (checkData($data, $dataNeeded)) {
+		//checks if requests needed are present and not empty
+		$dataNeeded = array("projectName", "skills", "longDescription", "shortDescription", "github", "date");
+		if (checkData($data, $dataNeeded)) {
 
-		$data["date"] = date("Y-m-d", strtotime($data["date"]));
+			$data["date"] = date("Y-m-d", strtotime($data["date"]));
 
-		$db = new pdodb;
-		$query = "INSERT INTO portfolioproject (Name, Skills, LongDescription, ShortDescription, Link, GitHub, Download, Date, Colour) VALUES (:projectName, :skills, :longDescription, :shortDescription, :link, :github, :download, :date, :colour);";
-		$bindings = array(":projectName" => $data["projectName"], ":skills" => $data["skills"], ":longDescription" => $data["longDescription"], ":shortDescription" => $data["shortDescription"], ":link" => $data["link"], ":github" => $data["github"], ":download" => $data["download"], ":date" => $data["date"], ":colour" => $data["colour"]);
-		$results = $db->query($query, $bindings);
+			$db = new pdodb;
+			$query = "INSERT INTO portfolioproject (Name, Skills, LongDescription, ShortDescription, Link, GitHub, Download, Date, Colour) VALUES (:projectName, :skills, :longDescription, :shortDescription, :link, :github, :download, :date, :colour);";
+			$bindings = array(":projectName" => $data["projectName"], ":skills" => $data["skills"], ":longDescription" => $data["longDescription"], ":shortDescription" => $data["shortDescription"], ":link" => $data["link"], ":github" => $data["github"], ":download" => $data["download"], ":date" => $data["date"], ":colour" => $data["colour"]);
+			$results = $db->query($query, $bindings);
 
-		//if add was ok
-		if ($results["count"] > 0) {
+			//if add was ok
+			if ($results["count"] > 0) {
 
-			$projectID = $db->lastInsertId();
-			$results = getProject($projectID);
+				$projectID = $db->lastInsertId();
+				$results = getProject($projectID);
 
-			$results["meta"]["ok"] = true;
-			$results["meta"]["status"] = 201;
-			$results["meta"]["message"] = "Created";
+				$results["meta"]["ok"] = true;
+				$results["meta"]["status"] = 201;
+				$results["meta"]["message"] = "Created";
 
-		} //else error adding project
-		else {
+			} //else error adding project
+			else {
 
-			//check if database provided any meta data if so problem with executing query
-			if (!isset($results["meta"])) {
-				$results["meta"]["ok"] = false;
+				//check if database provided any meta data if so problem with executing query
+				if (!isset($results["meta"])) {
+					$results["meta"]["ok"] = false;
+				}
 			}
-		}
 
-	} //else data was not provided
-	else {
-		$results["meta"] = dataNotProvided($dataNeeded);
+		} //else data was not provided
+		else {
+			$results["meta"] = dataNotProvided($dataNeeded);
+		}
+	}
+	else
+	{
+		$results = notAuthorised();
 	}
 
     return $results;
@@ -183,47 +203,53 @@ function addProject($data)
 //try to edit a project user has posted before
 function editProject($data)
 {
+	//checks if user is authored
+	if (isLoggedIn()) {
 
-	//checks if requests needed are present and not empty
-	$dataNeeded = array("projectID", "projectName", "skills", "longDescription", "shortDescription", "github", "date");
-	if (checkData($data, $dataNeeded)) {
+		//checks if requests needed are present and not empty
+		$dataNeeded = array("projectID", "projectName", "skills", "longDescription", "shortDescription", "github", "date");
+		if (checkData($data, $dataNeeded)) {
 
-		//Check the project trying to edit actually exists
-		$project = getProject($data["projectID"]);
-		if ($project["count"] > 0) {
+			//Check the project trying to edit actually exists
+			$project = getProject($data["projectID"]);
+			if ($project["count"] > 0) {
 
-			$data["date"] = date("Y-m-d", strtotime($data["date"]));
+				$data["date"] = date("Y-m-d", strtotime($data["date"]));
 
-			$db = new pdodb;
-			$query = "UPDATE portfolioproject SET Name = :projectName, Skills = :skills, LongDescription = :longDescription, Link = :link, ShortDescription = :shortDescription, GitHub = :github, Download = :download, Date = :date, Colour = :colour WHERE ID = :projectID;";
-			$bindings = array(":projectID" => $data["projectID"], ":projectName" => $data["projectName"], ":skills" => $data["skills"], ":longDescription" => $data["longDescription"], ":shortDescription" => $data["shortDescription"], ":link" => $data["link"], ":github" => $data["github"], ":download" => $data["download"], ":date" => $data["date"], ":colour" => $data["colour"]);
-			$results = $db->query($query, $bindings);
+				$db = new pdodb;
+				$query = "UPDATE portfolioproject SET Name = :projectName, Skills = :skills, LongDescription = :longDescription, Link = :link, ShortDescription = :shortDescription, GitHub = :github, Download = :download, Date = :date, Colour = :colour WHERE ID = :projectID;";
+				$bindings = array(":projectID" => $data["projectID"], ":projectName" => $data["projectName"], ":skills" => $data["skills"], ":longDescription" => $data["longDescription"], ":shortDescription" => $data["shortDescription"], ":link" => $data["link"], ":github" => $data["github"], ":download" => $data["download"], ":date" => $data["date"], ":colour" => $data["colour"]);
+				$results = $db->query($query, $bindings);
 
-			//if update was ok
-			if ($results["count"] > 0) {
-				$pictures = json_decode($data["pictures"]);
+				//if update was ok
+				if ($results["count"] > 0) {
+					$pictures = json_decode($data["pictures"]);
 
-				if (count($pictures) > 0) {
-					foreach ($pictures as $picture) {
-						$query = "UPDATE portfolioprojectimage SET Number = :Number WHERE ID = :ID;";
-						$bindings = array(":ID" => $picture->ID, ":Number" => $picture->Number);
-						$db->query($query, $bindings);
+					if (count($pictures) > 0) {
+						foreach ($pictures as $picture) {
+							$query = "UPDATE portfolioprojectimage SET Number = :Number WHERE ID = :ID;";
+							$bindings = array(":ID" => $picture->ID, ":Number" => $picture->Number);
+							$db->query($query, $bindings);
+						}
+					}
+
+					$results = getProject($data["projectID"]);
+					$results["meta"]["ok"] = true;
+				} //error updating project
+				else {
+					//check if database provided any meta data if so problem with executing query
+					if (!isset($project["meta"])) {
+						$results["meta"]["ok"] = false;
 					}
 				}
-
-				$results = getProject($data["projectID"]);
-				$results["meta"]["ok"] = true;
-			} //error updating project
-			else {
-				//check if database provided any meta data if so problem with executing query
-				if (!isset($project["meta"])) {
-					$results["meta"]["ok"] = false;
-				}
 			}
+		} //else data was not provided
+		else {
+			$results["meta"] = dataNotProvided($dataNeeded);
 		}
-	} //else data was not provided
+	}
 	else {
-		$results["meta"] = dataNotProvided($dataNeeded);
+		$results = notAuthorised();
 	}
 
     return $results;
@@ -233,40 +259,47 @@ function editProject($data)
 function deleteProject($data)
 {
 
-	//checks if requests needed are present and not empty
-	$dataNeeded = array("projectID");
-	if (checkData($data, $dataNeeded)) {
+	//checks if user is authored
+	if (isLoggedIn()) {
 
-		//Check the project trying to edit actually exists
-		$results = getProject($data["projectID"]);
-		if ($results["count"] > 0) {
+		//checks if requests needed are present and not empty
+		$dataNeeded = array("projectID");
+		if (checkData($data, $dataNeeded)) {
 
-			$db = new pdodb;
-			//Delete the images linked to project
-			$query = "DELETE FROM portfolioprojectimage WHERE ProjectID = :projectID;";
-			$bindings = array(":projectID" => $data["projectID"]);
-			$db->query($query, $bindings);
-
-			//finally delete the actual project
-			$query = "DELETE FROM portfolioproject WHERE ID = :projectID;";
-			$results = $db->query($query, $bindings);
-
-			//if deletion was ok
+			//Check the project trying to edit actually exists
+			$results = getProject($data["projectID"]);
 			if ($results["count"] > 0) {
-				$results["meta"]["ok"] = true;
 
-				$results["rows"]["projectID"] = $data["projectID"];
-			} //error deleting project
-			else {
-				//check if database provided any meta data if so problem with executing query
-				if (!isset($results["meta"])) {
-					$results["meta"]["ok"] = false;
+				$db = new pdodb;
+				//Delete the images linked to project
+				$query = "DELETE FROM portfolioprojectimage WHERE ProjectID = :projectID;";
+				$bindings = array(":projectID" => $data["projectID"]);
+				$db->query($query, $bindings);
+
+				//finally delete the actual project
+				$query = "DELETE FROM portfolioproject WHERE ID = :projectID;";
+				$results = $db->query($query, $bindings);
+
+				//if deletion was ok
+				if ($results["count"] > 0) {
+					$results["meta"]["ok"] = true;
+
+					$results["rows"]["projectID"] = $data["projectID"];
+				} //error deleting project
+				else {
+					//check if database provided any meta data if so problem with executing query
+					if (!isset($results["meta"])) {
+						$results["meta"]["ok"] = false;
+					}
 				}
 			}
+		} //else data was not provided
+		else {
+			$results["meta"] = dataNotProvided($dataNeeded);
 		}
-	} //else data was not provided
+	}
 	else {
-		$results["meta"] = dataNotProvided($dataNeeded);
+		$results = notAuthorised();
 	}
 
     return $results;
@@ -285,71 +318,78 @@ function getPictures($projectID)
 function addPicture($data)
 {
 
-	//checks if requests needed are present and not empty
-	$dataNeeded = array("projectID");
-	if (checkData($data, $dataNeeded) && isset($_FILES["picture"])) {
+	//checks if user is authored
+	if (isLoggedIn()) {
 
-		//Check the project trying to edit actually exists
-		$project = getProject($data["projectID"]);
-		if ($project["count"] > 0) {
+		//checks if requests needed are present and not empty
+		$dataNeeded = array("projectID");
+		if (checkData($data, $dataNeeded) && isset($_FILES["picture"])) {
 
-			//get the file type
-			$imageFileType = pathinfo(basename($_FILES["picture"]["name"]), PATHINFO_EXTENSION);
+			//Check the project trying to edit actually exists
+			$project = getProject($data["projectID"]);
+			if ($project["count"] > 0) {
 
-			//the directory to upload file
-			$directory = "/assets/images/projects/";
+				//get the file type
+				$imageFileType = pathinfo(basename($_FILES["picture"]["name"]), PATHINFO_EXTENSION);
 
-			//the full path for new file
-			$filename = date('YmdHis', time()) . mt_rand() . "." . $imageFileType;
-			$fileLocation = $directory . $filename;
-			$fullPath = $_SERVER['DOCUMENT_ROOT'] . $directory . $filename;
+				//the directory to upload file
+				$directory = "/assets/images/projects/";
 
-			//check if file is a actual image
-			$fileType = mime_content_type($_FILES["picture"]["tmp_name"]);
-			if ((strpos($fileType, 'image/') !== false)) {
+				//the full path for new file
+				$filename = date('YmdHis', time()) . mt_rand() . "." . $imageFileType;
+				$fileLocation = $directory . $filename;
+				$fullPath = $_SERVER['DOCUMENT_ROOT'] . $directory . $filename;
 
-				//try to upload file
-				if (move_uploaded_file($_FILES["picture"]["tmp_name"], $fullPath)) {
+				//check if file is a actual image
+				$fileType = mime_content_type($_FILES["picture"]["tmp_name"]);
+				if ((strpos($fileType, 'image/') !== false)) {
 
-					//update database with location of new picture
-					$db = new pdodb;
-					$query = "INSERT INTO portfolioprojectimage (File, ProjectID, Number) VALUES (:file, :projectID, 0);";
-					$bindings = array(":file" => $fileLocation, ":projectID" => $data["projectID"]);
-					$results = $db->query($query, $bindings);
+					//try to upload file
+					if (move_uploaded_file($_FILES["picture"]["tmp_name"], $fullPath)) {
 
-					//if update of user was ok
-					if ($results["count"] > 0) {
-
-						$query = "SELECT * FROM portfolioprojectimage WHERE File = :file AND ProjectID = :projectID;";
+						//update database with location of new picture
+						$db = new pdodb;
+						$query = "INSERT INTO portfolioprojectimage (File, ProjectID, Number) VALUES (:file, :projectID, 0);";
+						$bindings = array(":file" => $fileLocation, ":projectID" => $data["projectID"]);
 						$results = $db->query($query, $bindings);
 
-						$results["meta"]["ok"] = true;
-						$results["meta"]["status"] = 201;
-						$results["meta"]["message"] = "Created";
+						//if update of user was ok
+						if ($results["count"] > 0) {
+
+							$query = "SELECT * FROM portfolioprojectimage WHERE File = :file AND ProjectID = :projectID;";
+							$results = $db->query($query, $bindings);
+
+							$results["meta"]["ok"] = true;
+							$results["meta"]["status"] = 201;
+							$results["meta"]["message"] = "Created";
 
 
-					} //else error updating user
-					else {
-						//check if database provided any meta data if so problem with executing query
-						if (!isset($picture["meta"])) {
-							$results["meta"]["ok"] = false;
+						} //else error updating user
+						else {
+							//check if database provided any meta data if so problem with executing query
+							if (!isset($picture["meta"])) {
+								$results["meta"]["ok"] = false;
+							}
 						}
+					} //else problem uploading file to server
+					else {
+						$results["meta"]["feedback"] = "Sorry, there was an error uploading your file.";
 					}
-				} //else problem uploading file to server
+				} //else bad request as file uploaded is not a image
 				else {
-					$results["meta"]["feedback"] = "Sorry, there was an error uploading your file.";
+					$results["meta"]["status"] = 400;
+					$results["meta"]["message"] = "Bad Request";
+					$results["meta"]["feedback"] = "File is not an image.";
 				}
-			} //else bad request as file uploaded is not a image
-			else {
-				$results["meta"]["status"] = 400;
-				$results["meta"]["message"] = "Bad Request";
-				$results["meta"]["feedback"] = "File is not an image.";
 			}
+		} //else data was not provided
+		else {
+			array_push($dataNeeded, "pictureUploaded");
+			$results["meta"] = dataNotProvided($dataNeeded);
 		}
-	} //else data was not provided
+	}
 	else {
-		array_push($dataNeeded, "pictureUploaded");
-		$results["meta"] = dataNotProvided($dataNeeded);
+		$results = notAuthorised();
 	}
 
     $results["meta"]["files"] = $_FILES;
@@ -360,42 +400,50 @@ function addPicture($data)
 //Tries to delete a picture linked to a project
 function deletePicture($data)
 {
-	//checks if requests needed are present and not empty
-	$dataNeeded = array("projectID", "file");
-	if (checkData($data, $dataNeeded)) {
 
-		//Check the project trying to edit actually exists
-		$results = getProject($data["projectID"]);
-		if ($results["count"] > 0) {
+	//checks if user is authored
+	if (isLoggedIn()) {
 
-			//update database
-			$db = new pdodb;
-			$query = "DELETE FROM portfolioprojectimage WHERE ProjectID = :projectID AND File = :file;";
-			$bindings = array(":projectID" => $data["projectID"], ":file" => $data["file"]);
-			$results = $db->query($query, $bindings);
+		//checks if requests needed are present and not empty
+		$dataNeeded = array("projectID", "file");
+		if (checkData($data, $dataNeeded)) {
 
-			//if update was ok
+			//Check the project trying to edit actually exists
+			$results = getProject($data["projectID"]);
 			if ($results["count"] > 0) {
 
-				//checks if file exists to delete the picture
-				if (file_exists($data["file"])) {
-					unlink($data["file"]);
-				}
+				//update database
+				$db = new pdodb;
+				$query = "DELETE FROM portfolioprojectimage WHERE ProjectID = :projectID AND File = :file;";
+				$bindings = array(":projectID" => $data["projectID"], ":file" => $data["file"]);
+				$results = $db->query($query, $bindings);
 
-				$results["meta"]["ok"] = true;
-				$results["rows"]["file"] = $data["file"];
+				//if update was ok
+				if ($results["count"] > 0) {
 
-			} //else error updating
-			else {
-				//check if database provided any meta data if so problem with executing query
-				if (!isset($results["meta"])) {
-					$results["meta"]["ok"] = false;
+					//checks if file exists to delete the picture
+					if (file_exists($data["file"])) {
+						unlink($data["file"]);
+					}
+
+					$results["meta"]["ok"] = true;
+					$results["rows"]["file"] = $data["file"];
+
+				} //else error updating
+				else {
+					//check if database provided any meta data if so problem with executing query
+					if (!isset($results["meta"])) {
+						$results["meta"]["ok"] = false;
+					}
 				}
 			}
+		} //else data was not provided
+		else {
+			$results["meta"] = dataNotProvided($dataNeeded);
 		}
-	} //else data was not provided
+	}
 	else {
-		$results["meta"] = dataNotProvided($dataNeeded);
+		$results = notAuthorised();
 	}
 
     return $results;
